requirements.txt
langchain
requests
beautifulsoup4
tiktoken
langchain-community
langchain-groq
python-dotenv